+++ src/api/processors/trim.cpp
@@ -4,6 +4,8 @@
 
 namespace weserv::api::processors {
 
+using parsers::Color;
+
 VImage Trim::process(const VImage &image) const {
     auto threshold = query_->get_if<int>(
         "trim",
@@ -22,14 +24,36 @@ VImage Trim::process(const VImage &image) const {
         return image;
     }
 
-    // Find the value of the pixel at (0, 0), `find_trim` search for all pixels
-    // significantly different from this
-    auto background = image.extract_area(0, 0, 1, 1);
+    std::vector<double> trim_background;
+    if (query_->exists("tbg")) {
+        trim_background = query_->get<Color>("tbg").to_rgba();
 
-    // Note: If the image has alpha, we'll need to flatten before `getpoint`
-    // to get a correct background value
-    if (image.has_alpha()) {
-        background = background.flatten();
+        if (utils::is_16_bit(image.interpretation())) {
+            for (auto &i : trim_background) {
+                i *= 256.0;
+            }
+        }
+
+        if (image.bands() < 3) {
+            // Convert sRGB to greyscale
+            trim_background = {0.2126 * trim_background[0] +
+                               0.7152 * trim_background[1] +
+                               0.0722 * trim_background[2]};
+        } else {
+            auto bands = image.has_alpha() ? image.bands() - 1 : image.bands();
+            trim_background.resize(bands, trim_background[3]);
+        }
+    } else {
+        // Top-left pixel provides the default background color if none is given
+        auto background = image.extract_area(0, 0, 1, 1);
+
+        // Note: If the image has alpha, we'll need to flatten before `getpoint`
+        // to get a correct background value
+        if (image.has_alpha()) {
+            background = background.flatten();
+        }
+
+        trim_background = background(0, 0);
     }
 
     // Scale up 8-bit values to match 16-bit input image
@@ -41,7 +65,7 @@ VImage Trim::process(const VImage &image) const {
     left = image.find_trim(&top, &width, &height,
                            VImage::option()
                                ->set("threshold", threshold)
-                               ->set("background", background(0, 0)));
+                               ->set("background", trim_background));
 
     // Sanity check, this usually happens when a high tolerance is specified
     if (width == 0 || height == 0) {