+++ test/api/processors/unit-trim.cpp
@@ -23,10 +23,25 @@ TEST_CASE("trim", "[trim]") {
         CHECK_THAT(image, is_similar_image(expected_image));
     }
 
+    SECTION("background") {
+        auto test_image = fixtures->input_png_with_grey_alpha;
+        auto expected_image =
+            fixtures->expected_dir + "/grey-8bit-alpha-trim.png";
+        auto params = "tbg=0FFF&trim=25";
+
+        VImage image = process_file<VImage>(test_image, params);
+
+        CHECK(image.width() == 282);
+        CHECK(image.height() == 238);
+        CHECK(image.has_alpha());
+
+        CHECK_THAT(image, is_similar_image(expected_image));
+    }
+
     SECTION("16bit with transparency") {
         auto test_image = fixtures->input_png_with_transparency_16bit;
         auto expected_image = fixtures->expected_dir + "/trim-16bit-rgba.png";
-        auto params = "w=32&h=32&fit=cover&trim=10";
+        auto params = "w=32&h=32&fit=cover&tbg=0FFF&trim=10";
 
         VImage image = process_file<VImage>(test_image, params);